Refactor PostsList
